requirements.txt
click==6.6
Flask==0.11.1
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
omdb==0.7.0
pygal==2.3.0
PyYAML==3.12
requests==2.11.1
Werkzeug==0.11.11
